Changes from 2.50-se to 2.51-SE path_95x no updated

+++ otxserver2/path_97x/sources/iologindata.cpp
@@ -112,13 +112,7 @@ void IOLoginData::loadCharacters(Account& account)
 		account.charList.push_back(result->getDataString("name"));
 #else
 		std::string name = result->getDataString("name");
-		if(hasCustomFlag(PlayerCustomFlag_GamemasterPrivileges, result->getDataInt("id")))
-		{
-			uint16_t hax = 0;
-			for(GameServersMap::const_iterator it = GameServers::getInstance()->getFirstServer(); it != GameServers::getInstance()->getLastServer(); ++it, ++hax)
-				account.charList[name + asString(hax)] = Character(name, it->second, -1);
-		}
-		else if(GameServer* srv = GameServers::getInstance()->getServerById(result->getDataInt("world_id")))
+		if(GameServer* srv = GameServers::getInstance()->getServerById(result->getDataInt("world_id")))
 			account.charList[name] = Character(name, srv, result->getDataInt("online"));
 		else
 			std::clog << "[Warning - IOLoginData::loadAccount] Invalid server for player '" << name << "'." << std::endl;
@@ -740,30 +734,62 @@ bool IOLoginData::loadPlayer(Player* player, const std::string& name, bool preLo
 	if((result = db->storeQuery(query.str())))
 	{
 		loadItems(itemMap, result);
-		for(ItemMap::reverse_iterator rit = itemMap.rbegin(); rit != itemMap.rend(); ++rit)
+
+		ItemMap::reverse_iterator it;
+		ItemMap::iterator it2;
+		for(it = itemMap.rbegin(); it != itemMap.rend(); ++it)
 		{
-			Item* item = rit->second.first;
-			int32_t pid = rit->second.second;
+			Item* item = it->second.first;  
+			int32_t pid = it->second.second;
 			if(pid >= 0 && pid < 100)
 			{
-				if(Container* c = item->getContainer())
-				{
-					if(Depot* depot = c->getDepot())
-						player->addDepot(depot, pid);
-					else
-						std::clog << "[Error - IOLoginData::loadPlayer] Cannot load depot " << pid << " for player " << name << std::endl;
-				}
-				else
-					std::clog << "[Error - IOLoginData::loadPlayer] Cannot load depot " << pid << " for player " << name << std::endl;
+				DepotChest* depotChest = player->getDepotChest(pid, true);
+				if(depotChest)
+					depotChest->__internalAddThing(item);
 			}
-			else if((it = itemMap.find(pid)) != itemMap.end())
+			else
 			{
-				if(Container* container = it->second.first->getContainer())
+				it2 = itemMap.find(pid);
+				if(it2 == itemMap.end())
+					continue;
+
+				Container* container = it2->second.first->getContainer();
+				if(container)
 					container->__internalAddThing(item);
 			}
 		}
+		itemMap.clear();
+		result->free();
+	}
 
-		player->updateDepots();
+	//load inbox items
+	query.str("");
+	query << "SELECT `pid`, `sid`, `itemtype`, `count`, `attributes` FROM `player_inboxitems` WHERE `player_id` = " << player->getGUID() << " ORDER BY `sid` DESC;";
+	if((result = db->storeQuery(query.str())))
+	{
+		loadItems(itemMap, result);
+
+		ItemMap::reverse_iterator it;
+		ItemMap::iterator it2;
+		for(it = itemMap.rbegin(); it != itemMap.rend(); ++it)
+		{
+			Item* item = it->second.first;
+			int32_t pid = it->second.second;
+			if(pid >= 0 && pid < 100)
+			{
+				player->getInbox()->__internalAddThing(item);
+			}
+			else
+			{
+				it2 = itemMap.find(pid);
+				if(it2 == itemMap.end())
+					continue;
+
+				Container* container = it2->second.first->getContainer();
+				if(container)
+					container->__internalAddThing(item);
+			}
+		}
 		itemMap.clear();
 		result->free();
 	}
@@ -795,7 +821,7 @@ bool IOLoginData::loadPlayer(Player* player, const std::string& name, bool preLo
 			std::string description = result->getDataString("description");
 			bool notify = !!result->getDataInt("notify");
 			if(storeNameByGuid(vid))
-				player->addVIP(vid, "", description, icon, notify, false, true);
+				player->addVIP(vid, "", description, icon, notify, VIPSTATUS_ONLINE, true);
 		}
 		while(result->next());
 		result->free();
@@ -1043,38 +1069,40 @@ bool IOLoginData::savePlayer(Player* player, bool preSave/* = true*/, bool shall
 	if(!saveItems(player, itemList, stmt))
 		return false;
 
-	itemList.clear();
-	//save depot items
-	//std::stringstream ss;
-	for(DepotMap::iterator it = player->depots.begin(); it != player->depots.end(); ++it)
+	if(player->depotChange)
 	{
-		/*if(it->second.second)
-		{
-			it->second.second = false;
-			ss << it->first << ",";*/
-			itemList.push_back(itemBlock(it->first, it->second.first));
-		//}
-	}
-
-	/*std::string s = ss.str();
-	size_t size = s.length();
-	if(size > 0)
-	{*/
-
+		//save depot items
 		query.str("");
-		query << "DELETE FROM `player_depotitems` WHERE `player_id` = " << player->getGUID();// << " AND `pid` IN (" << s.substr(0, --size) << ")";
+		query << "DELETE FROM `player_depotitems` WHERE `player_id` = " << player->getGUID() << ";";
 		if(!db->query(query.str()))
 			return false;
 
-		if(itemList.size())
+		stmt.setQuery("INSERT INTO `player_depotitems` (`player_id`, `pid`, `sid`, `itemtype`, `count`, `attributes`) VALUES ");
+		itemList.clear();
+		for(DepotMap::iterator it = player->depotChests.begin(); it != player->depotChests.end() ;++it)
 		{
-			stmt.setQuery("INSERT INTO `player_depotitems` (`player_id`, `pid`, `sid`, `itemtype`, `count`, `attributes`) VALUES ");
-			if(!saveItems(player, itemList, stmt))
-				return false;
-
-			itemList.clear();
+			DepotChest* depotChest = it->second;
+			for(ItemList::const_iterator iit = depotChest->getItems(), end = depotChest->getEnd(); iit != end; ++iit)
+				itemList.push_back(itemBlock(it->first, *iit));
 		}
-	//}
+
+		if(!saveItems(player, itemList, stmt))
+			return false;
+	}
+
+	//save inbox items
+	query.str("");
+	query << "DELETE FROM `player_inboxitems` WHERE `player_id` = " << player->getGUID() << ";";
+	if(!db->query(query.str()))
+		return false;
+
+	stmt.setQuery("INSERT INTO `player_inboxitems` (`player_id`, `pid`, `sid`, `itemtype`, `count`, `attributes`) VALUES ");
+	itemList.clear();
+	for(ItemList::const_iterator it = player->getInbox()->getItems(), end = player->getInbox()->getEnd(); it != end; ++it)
+		itemList.push_back(itemBlock(0, *it));
+
+	if(!saveItems(player, itemList, stmt))
+		return false;
 
 	query.str("");
 	query << "DELETE FROM `player_storage` WHERE `player_id` = " << player->getGUID();
@@ -1304,21 +1332,13 @@ bool IOLoginData::playerDeath(Player* _player, const DeathList& dl)
 	return trans.commit();
 }
 
-bool IOLoginData::playerMail(Creature* actor, std::string name, uint32_t townId, Item* item)
+bool IOLoginData::playerMail(Creature* actor, std::string name, Item* item)
 {
 	Player* player = g_game.getPlayerByNameEx(name);
 	if(!player)
 		return false;
 
-	if(!townId)
-		townId = player->getTown();
-
-	Depot* depot = player->getDepot(townId, true);
-	Container* inbox = NULL;
-	if(!(inbox = depot->getInbox()))
-		inbox = depot;
-
-	if(!inbox || g_game.internalMoveItem(actor, item->getParent(), inbox, INDEX_WHEREEVER,
+	if(g_game.internalMoveItem(actor, item->getParent(), player->getInbox(), INDEX_WHEREEVER,
 		item, item->getItemCount(), NULL, FLAG_NOLIMIT) != RET_NOERROR)
 	{
 		if(player->isVirtual())
@@ -1328,7 +1348,7 @@ bool IOLoginData::playerMail(Creature* actor, std::string name, uint32_t townId,
 	}
 
 	g_game.transformItem(item, ITEM_PARCEL_STAMPED);
-	bool result = true, opened = player->getContainerID(depot) != -1;
+	bool result = true, opened = player->getContainerID(player->getInbox()) != -1;
 
 	Player* tmp = NULL;
 	if(actor)
